Провел рефакторинг - AboutView компонент
