+++ src/views/AboutUsView.vue
@@ -1,147 +1,60 @@
 <template>
     <div class="container">
-
         <Dock :model="items" :position="position">
             <template #itemicon="{ item }">
-                <img class="dock__icon" v-tooltip.top="item.label" :alt="item.label" :src="item.icon" style="width: 90%" />
+                <DockIcon :item="item" />
             </template>
         </Dock>
 
         <div class="left__section">
-
             <div class="left__section_container">
-
-                <div>
-                    <glitch-text :text="'Добро пожаловать в Abstergo'"></glitch-text>
-                </div>
-
-                <div class="left__section_container-aboutUs_text">
-                    <h1 class="aboutUs-text">
-                        Мы — команда экспертов, создающая инновационные решения для вашего бизнеса. Разрабатываем
-                        мобильные приложения, десктопные и веб-приложения, серверные решения и многое другое. Дайте нам
-                        шанс превратить ваши идеи в реальность и вывести ваш проект на новый уровень
-                    </h1>
-                </div>
-
+                <WelcomeText text="Добро пожаловать в Abstergo" />
+                <AboutUsText class="mt-6" :text="aboutUsDescription" />
                 <div class="left__section_container-button">
                     <primary-button :color-text-hex="'#ffffff'" :color-hex="'#ffffff'" :outlined="true"
-                        :text="'Посмотреть услуги'" @click="goOurServices()"></primary-button>
+                        :text="'Посмотреть услуги'" @click="goOurServices()" />
                 </div>
-
-                <div class="bottom__container">
-                    <div class="phone__number">+7 999 600 22 11</div>
-                    <div class="pascal__text">100 Main St. New York NY 10002. USA  - Сайт написан на vue3 + typescript</div>
-                </div>
-
+                <ContactInfo phone="+7 999 600 22 11"
+                    address="100 Main St. New York NY 10002. USA - Сайт написан на vue3 + typescript" />
             </div>
-
         </div>
 
         <div class="right__section">
-            <div v-for="it in baseServicesList" :key="it.id" class="blurred-box service__item"
-                :style="{ ...blurBoxStyle, ...{ marginTop: '10px', display: 'flex' } }">
-                <div class="service__inner_container">
-
-                    <div class="top__container">
-                        <img :src="it.icon" class="service__icon">
-                        <span class="service__label">{{ it.label }}</span>
-                    </div>
-
-                    <div class="down__container">
-                        {{ it.desc }}
-                    </div>
-
-
-                </div>
+            <div v-for="it in baseServicesList" :key="it.id">
+                <ServiceItem :icon="it.icon" :label="it.label" :desc="it.desc" />
             </div>
         </div>
-
     </div>
 </template>
+
 <script setup>
-import GlitchText from '@/components/ui/Effects/GlitchText.vue';
+import WelcomeText from '@/components/AboutView/WelcomeText.vue';
+import AboutUsText from '@/components/AboutView/AboutUsText.vue';
+import ContactInfo from '@/components/AboutView/ContactInfo.vue';
+import ServiceItem from '@/components/AboutView/ServiceItem.vue';
+import DockIcon from '@/components/AboutView/DockIcon.vue';
+
 import PrimaryButton from '@/components/ui/Buttons/PrimaryButton.vue';
-
 import Dock from 'primevue/dock';
-
-import { useBlurBox } from '@/components/ui/Effects/blurred.box';
-
-import botIcon from '@/assets/icons/services-icons/develop-bots.png';
-import mobIcon from '@/assets/icons/services-icons/develop-mob.png';
-import serverIcon from '@/assets/icons/services-icons/servers-icon.png';
-
-import { ref } from 'vue';
 import { useRouter } from 'vue-router';
+import { useAboutView } from '@/composables/views/AboutView/useAboutView';
+import { ref } from 'vue';
 
+// Инициализация
 const router = useRouter();
+const { baseServicesList, items, position } = useAboutView();
 
-const { blurBoxStyle } = useBlurBox({ width: '400px', height: '230px' });
-
-// Init data
-
-const baseServicesList = ref([
-    {
-        id: 0,
-        right: false,
-        label: 'Мобильные приложения',
-        desc: 'Разработка мобильных приложений для iOS и Android. Наша команда обеспечивает полный цикл услуг: от концепции и дизайна до разработки и поддержки',
-        // icon: 'develop-mob.png',
-        icon: mobIcon,
-    },
-    {
-        id: 1,
-        right: true,
-        label: 'Боты, парсеры, плагины',
-        desc: 'Разработке ботов, парсеров и плагинов. Наша команда создает эффективные решения для автоматизации процессов, извлечения данных и интеграции с различными платформами',
-        // icon: 'develop-bots.png',
-        icon: botIcon,
-    },
-    {
-        id: 2,
-        right: false,
-        label: 'Сервера, Базы данных',
-        desc: 'Разработка и управлении серверами и базами данных. Мы предоставляем услуги по настройке, мониторингу и оптимизации серверной инфраструктуры',
-        // icon: 'servers-icon.png',
-        icon: serverIcon,
-    },
-])
-
-// Dock features
-import telegramIcon from '@/assets/icons/networks/telegram-icon.png';
-import githubIcon from '@/assets/icons/networks/github-icon.png';
-import twitterIcon from '@/assets/icons/networks/twitter.png';
-
-const items = ref([
-    {
-        label: 'App Store',
-        icon: twitterIcon
-    },
-    {
-        label: 'Photos',
-        icon: telegramIcon
-    },
-    {
-        label: 'Trash',
-        icon: githubIcon
-    }
-]);
-const position = ref('right');
-
-// Асинхронное получение динамичного пути по спецификации vite сборщика
-async function getIconPath(icon) {
-    try {
-        const image = await import(`@/assets/icons/services-icons/${icon}`);
-        return image.default; // Возвращаем путь к изображению
-    } catch (error) {
-        console.error('Ошибка загрузки изображения:', error);
-        return ''; // Возвращаем пустую строку в случае ошибки
-    }
-}
+const aboutUsDescription = ref(
+    `
+    Мы — команда экспертов, создающая инновационные решения для вашего бизнеса. Разрабатываем
+    мобильные приложения, десктопные и веб-приложения, серверные решения и многое другое. Дайте нам
+    шанс превратить ваши идеи в реальность и вывести ваш проект на новый уровень
+    `
+);
 
 // Переход на страницу сервесов
-function goOurServices() {
-    router.push({ name: 'our-services' })
-}
+const goOurServices = () => router.push({ name: 'our-services' });
+
 </script>
 
 <style scoped>
@@ -193,31 +106,6 @@ function goOurServices() {
     margin-top: 20px;
 }
 
-/* Нижняя часть слева */
-.bottom__container {
-    width: 100%;
-    height: fit-content;
-    /* margin-top: auto; */
-    display: flex;
-    align-items: center;
-    position: absolute;
-    bottom: 0;
-}
-
-.bottom__container .phone__number {
-    color: white;
-    padding-left: 20px;
-    border-left: 1px solid #ffffff;
-    font-size: 15px;
-}
-
-.bottom__container .pascal__text {
-    color: #bcd4e7;
-    font-size: 12px;
-    margin-left: 20px;
-}
-
-
 /* RIGHT SIDE */
 .right__section {
     position: relative;
@@ -225,48 +113,6 @@ function goOurServices() {
     height: 100%;
     display: flex;
     flex-direction: column;
-}
-
-.service__inner_container {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    padding: 20px;
-}
-
-.service__inner_container .top__container {
-    width: 100%;
-    height: fit-content;
-    padding: 5px;
-    display: flex;
-    align-items: center;
-}
-
-.top__container .service__icon {
-    width: 50px;
-    height: 50px;
-    object-fit: contain;
-    border-radius: 10rem;
-    margin-right: 15px;
-}
-
-.top__container .service__label {
-    font-size: 15px;
-    font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
-    font-weight: 500;
-    color: white;
-}
-
-.service__inner_container .down__container {
-    width: 100%;
-    height: fit-content;
-    margin-top: 10px;
-    padding: 5px;
-    display: flex;
-    text-align: start;
-    text-wrap: wrap;
-    color: rgb(219, 219, 219);
 }
 
 .inner__container {
@@ -332,7 +178,4 @@ function goOurServices() {
     }
 }
 
-
-
-/* Effects */
 </style>
